requirements.txt
PyYAML==6.0.1
numpy==1.26.2
pandas==1.3.5
torch==2.3.0
torchvision==0.18.0
pytorch-ood==0.1.6
scikit-learn==1.3.2
scikit-image==0.22.0
scipy==1.11.4
kornia==0.7.0
statsmodels==0.14.0
matplotlib==3.8.2
seaborn==0.13.0
zennit==0.5.1
tqdm==4.66.1
adjustText==1.0.4
ipykernel==6.26.0
nbconvert==7.11.0
pandocfilters==1.5.1
timm==0.6.13